Reliability update
When having many subscribe/unsubscribes in a short time, a reliability problem became apparent:
The next subscription had been send before the last one received its confirmation from the broker, resulting in an endless reconnect as the received pid did not match the last pid sent to the broker.
Solution: Added an additional Lock for subscribe/unsubscribe to wait until the confirmation is received or try again.