Optimize memory usage
Optimize the memory usage of `.as_read()` and `.as_write()` functions. This is
done primarily be using memory view to avoid making copies during slicing of
the byte array. But also by changing the way memory is allocated when reading
from the socket to prevent memory reallocation. This reallocation could lead
to fragmentation of memory, preventing optimal use of it.