Cleanup and support for properties on received messages
This commit adds support for properties on received messages. This is a
breaking change for anyone who wants to use MQTTv5, as the handlers need
to be updated to take a new argument. As discussed in #127 this breaking
change is acceptable as there are no changes for anyone using MQTT v3.1.1.

Additionally, there is some minor cleanup in this commit, as some hex
numbers had minor changes in them. This commit reverts those changes.